requirements.txt
python-docx==1.1.0
requests==2.31.0
google-api-python-client==2.100.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dotenv==1.1.0
flask==3.1.1
psycopg2-binary==2.9.10
redis==4.5.0 